refactor: extract service layer
